requirements.txt
numpy==1.18.1
matplotlib==3.1.3
Pillow==8.0.0
